refactor(project-pattern): merge fix updating with new PR
